Feature/api error handling

+++ src/main/java/ca/sheridancollege/bakerdam/authsystemapi/service/UserServiceImpl.java
@@ -2,6 +2,8 @@ package ca.sheridancollege.bakerdam.authsystemapi.service;
 
 import ca.sheridancollege.bakerdam.authsystemapi.dto.request.CreateUserRequest;
 import ca.sheridancollege.bakerdam.authsystemapi.entity.UserEntity;
+import ca.sheridancollege.bakerdam.authsystemapi.exception.EmailAlreadyExistsException;
+import ca.sheridancollege.bakerdam.authsystemapi.exception.UserNotFoundException;
 import ca.sheridancollege.bakerdam.authsystemapi.repository.UserRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -20,6 +22,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserEntity saveUser(CreateUserRequest request) {
+        if (userRepository.existsByEmail(request.getEmail())) {
+            throw new EmailAlreadyExistsException(request.getEmail());
+        }
+
         UserEntity user = new UserEntity();
         user.setEmail(request.getEmail());
         user.setPassword(passwordEncoder.encode(request.getPassword()));
@@ -29,13 +35,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
+
         userRepository.deleteById(id);
     }
 
     @Override
     public UserEntity findUserById(Long id) {
         return userRepository.findById(id)
-                            .orElseThrow(() -> new RuntimeException("User not found"));
+                            .orElseThrow(() -> new UserNotFoundException(id));
     }
 
     @Override
@@ -46,7 +56,12 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserEntity updateUser(Long id, String email) { // Updates email only
         UserEntity existingUser = userRepository.findById(id)
-                                        .orElseThrow(() -> new RuntimeException("User not found"));
+                                        .orElseThrow(() -> new UserNotFoundException(id));
+
+        // If the user is changing their email and the email already exists
+        if (!existingUser.getEmail().equals(email) && userRepository.existsByEmail(email)) {
+            throw new EmailAlreadyExistsException(email);
+        }
         existingUser.setEmail(email);
 
         return userRepository.save(existingUser);
@@ -55,7 +70,7 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserEntity updatePassword(Long id, String password) {
         UserEntity user = userRepository.findById(id)
-                                        .orElseThrow(() -> new RuntimeException("User not found"));
+                                        .orElseThrow(() -> new UserNotFoundException(id));
 
         user.setPassword(passwordEncoder.encode(password));
